[Playground] [Backend] removed cloud storage client
