+++ playground/infrastructure/cd_helper.py
@@ -27,7 +27,7 @@ from config import Origin
 from datastore_client import DatastoreClient
 from grpc_client import GRPCClient
 from helper import Example, get_statuses
-from storage_client import StorageClient
+from repository import set_dataset_path_for_examples
 
 
 class CDHelper:
@@ -51,8 +51,7 @@ class CDHelper:
         """
         single_file_examples = list(filter(
             lambda example: example.tag.multifile is False, examples))
-        storage_client = StorageClient()
-        storage_client.set_dataset_path_for_examples(examples)
+        set_dataset_path_for_examples(single_file_examples)
         logging.info("Start of executing only single-file Playground examples ...")
         asyncio.run(self._get_outputs(single_file_examples))
         logging.info("Finish of executing single-file Playground examples")
@@ -80,6 +79,7 @@ class CDHelper:
         Args:
             examples: beam examples that should be run
         """
+
         async def _populate_fields(example: Example):
             try:
                 example.compile_output = await client.get_compile_output(example.pipeline_id)
@@ -94,7 +94,6 @@ class CDHelper:
 
         async with GRPCClient() as client:
             await get_statuses(client,
-                examples)  # run examples code and wait until all are executed
+                               examples)  # run examples code and wait until all are executed
             tasks = [_populate_fields(example) for example in examples]
             await asyncio.gather(*tasks)
-
