+++ playground/infrastructure/proxy/allow_list.py
@@ -20,14 +20,12 @@ ALLOWED_LIST = [
     "127.0.0.1",
     "logging.googleapis.com",
     "datastore.googleapis.com",
-    "oauth2.googleapis.com",
-    "storage.googleapis.com"
+    "oauth2.googleapis.com"
 ]
 
 # ALLOWED_BUCKET_LIST contains all public Google Cloud Storage buckets
 # that are allowed to make requests from Beam Playground
 ALLOWED_BUCKET_LIST = [
     "dataflow-samples",
-    "beam-samples",
-    "apache-beam-samples"
+    "beam-samples"
 ]
