+++ playground/infrastructure/requirements.txt
@@ -26,5 +26,4 @@ PyYAML==6.0
 tqdm~=4.62.3
 google-cloud-datastore==2.7.1
 sonora==0.2.2
-google-cloud-storage==1.43.0
 dataclasses-json==0.5.7
